finish refactor of Event
