+++ src/EventImpl.hpp
@@ -2,24 +2,8 @@
 
 
 #include <cassert>
-#include <Core.hpp>
-#include <functional>
 
-
-using CallBack = std::function<void()>;
-
-
-struct Event
-{
-    int fd;
-    EventType type;
-    CallBack  cb;
-
-    bool operator==(const Event& e) const
-    {
-        return e.fd == fd && e.type == type;
-    }
-};
+#include "Event.hpp"
 
 
 void AddEvent(int fd, EventType event_type, CallBack cb);
